requirements.txt
numpy
pandas
tqdm
jiwer
Pillow
scikit-learn
pyyaml
wandb
torch==2.4.0
transformers==4.51.3
datasets==3.0.1
evaluate==0.4.3
accelerate==0.34.2
bitsandbytes==0.42.0
peft==0.13.0
trl==0.11.1
